error in event destroy fixed

+++ app/controllers/exams_controller.rb
@@ -84,7 +84,7 @@ class ExamsController < ApplicationController
     @exam = Exam.find(params[:id])
     event = @exam.event
     if  @exam.destroy
-      event.event_series.destroy
+      event.event_series.destroy if event
       render :template => 'exams/destroy_success'
     else
       render :template => 'exams/destroy_failure'
